requirements.txt
streamlit==1.31.0
pandas==2.1.4
plotly==5.18.0
openpyxl==3.1.2
geopy==2.4.1
pyswisseph==2.10.3.2
matplotlib==3.8.2
